Coverted to JS modules and updated description processing
