+++ index.html
@@ -17,6 +17,6 @@
       <hr />
       <div id="filePreview"></div>
     </div>
-    <script type="text/javascript" src="app.js"></script>
+    <script type="module" src="/src/app.js"></script>
   </body>
 </html>
